+++ scripts/git-subtree/openCC/pull.ps1
@@ -1,4 +1,4 @@
 git subtree pull `
-    -P workspace/plugins/chinese-convert `
-    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-chinese-convert.git `
+    -P workspace/plugins/openCC `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-openCC.git `
     dev